Add Trap Models and scripts to traps

+++ Assets/Zombee/Scripts/GameManager.cs
@@ -22,8 +22,8 @@ public enum TrapType
 public struct TrapDef
 {
     public TrapType TrapType;
-    public int TimeToActive;
-    public int TimeActive;
+    public float TimeToActive;
+    public float TimeActive;
     public int Damage;
 }
